Introduce RemoteIdentifierSupplier for customizing remote identifiers
