+++ plugin/trino-base-jdbc/src/main/java/io/trino/plugin/jdbc/mapping/DefaultIdentifierMapping.java
@@ -13,21 +13,24 @@
  */
 package io.trino.plugin.jdbc.mapping;
 
-import io.trino.spi.TrinoException;
+import com.google.inject.Inject;
 import io.trino.spi.security.ConnectorIdentity;
 
 import java.sql.Connection;
-import java.sql.DatabaseMetaData;
-import java.sql.SQLException;
 
-import static io.trino.plugin.jdbc.JdbcErrorCode.JDBC_ERROR;
 import static java.util.Locale.ENGLISH;
+import static java.util.Objects.requireNonNull;
 
 public class DefaultIdentifierMapping
         implements IdentifierMapping
 {
-    // Caching this on a field is LazyConnectorFactory friendly
-    private Boolean storesUpperCaseIdentifiers;
+    private final RemoteIdentifierSupplier remoteIdentifierSupplier;
+
+    @Inject
+    public DefaultIdentifierMapping(RemoteIdentifierSupplier remoteIdentifierSupplier)
+    {
+        this.remoteIdentifierSupplier = requireNonNull(remoteIdentifierSupplier, "remoteIdentifierSupplier is null");
+    }
 
     @Override
     public String fromRemoteSchemaName(String remoteSchemaName)
@@ -67,24 +70,6 @@ public class DefaultIdentifierMapping
 
     private String toRemoteIdentifier(Connection connection, String identifier)
     {
-        if (storesUpperCaseIdentifiers(connection)) {
-            return identifier.toUpperCase(ENGLISH);
-        }
-        return identifier;
-    }
-
-    private boolean storesUpperCaseIdentifiers(Connection connection)
-    {
-        if (storesUpperCaseIdentifiers != null) {
-            return storesUpperCaseIdentifiers;
-        }
-        try {
-            DatabaseMetaData metadata = connection.getMetaData();
-            storesUpperCaseIdentifiers = metadata.storesUpperCaseIdentifiers();
-            return storesUpperCaseIdentifiers;
-        }
-        catch (SQLException e) {
-            throw new TrinoException(JDBC_ERROR, e);
-        }
+        return remoteIdentifierSupplier.toRemoteIdentifier(connection, identifier);
     }
 }
